backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
alembic==1.15.2
pydantic==2.10.3
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
redis==6.0.0
celery==5.5.2
psycopg2-binary==2.9.10
faiss-gpu==1.8.0
python-dotenv==1.1.0
ffmpeg-python==0.2.0
httpx==0.28.1
# 如果需要部署到 AWS/阿里云等云服务，取消注释以下行
# boto3>=1.26.0
# 向量化模型依赖
sentence-transformers==4.1.0
transformers==4.51.3
# 语音识别
faster-whisper==1.1.1
# 额外的视频处理依赖
imageio-ffmpeg==0.6.0
